update azure-mongo example
